Build separate float16 versions of the shaders
In order to be able to compare regular and float16 the demo needs to be
able to select the appropriate version of the shader to load when it
starts up. This change prepares for that by adding separate float16
versions of the shaders and the Makefile infrastructure to build them.